refactor: break apart menu
