+++ src/Menu/FeatureListRenderer.h
@@ -1,0 +1,79 @@
+#pragma once
+
+#include <functional>
+#include <map>
+#include <string>
+#include <variant>
+#include <vector>
+
+struct Feature;
+
+class FeatureListRenderer
+{
+public:
+	struct BuiltInMenu
+	{
+		std::string name;
+		std::function<void()> func;
+	};
+
+	struct CategoryHeader
+	{
+		std::string name;
+	};
+
+	using MenuFuncInfo = std::variant<BuiltInMenu, std::string, CategoryHeader, Feature*>;
+
+	static void RenderFeatureList(
+		float footerHeight,
+		size_t& selectedMenu,
+		std::string& featureSearch,
+		std::string& pendingFeatureSelection,
+		std::map<std::string, bool>& categoryExpansionStates,
+		const std::function<void()>& drawGeneralSettings,
+		const std::function<void()>& drawAdvancedSettings,
+		const std::function<void()>& drawDisplaySettings);
+
+private:
+	struct ListMenuVisitor
+	{
+		size_t listId;
+		size_t& selectedMenuRef;
+		std::map<std::string, bool>& categoryExpansionStates;
+
+		void operator()(const BuiltInMenu& menu);
+		void operator()(const std::string& label);
+		void operator()(const CategoryHeader& header);
+		void operator()(Feature* feat);
+	};
+
+	struct DrawMenuVisitor
+	{
+		void operator()(const BuiltInMenu& menu);
+		void operator()(const std::string&);
+		void operator()(const CategoryHeader&);
+		void operator()(Feature* feat);
+	};
+
+	static std::vector<MenuFuncInfo> BuildMenuList(
+		const std::string& featureSearch,
+		std::map<std::string, bool>& categoryExpansionStates,
+		const std::function<void()>& drawGeneralSettings,
+		const std::function<void()>& drawAdvancedSettings,
+		const std::function<void()>& drawDisplaySettings);
+
+	static void HandlePendingFeatureSelection(
+		std::string& pendingFeatureSelection,
+		const std::vector<MenuFuncInfo>& menuList,
+		size_t& selectedMenu);
+
+	static void RenderLeftColumn(
+		const std::vector<MenuFuncInfo>& menuList,
+		size_t& selectedMenu,
+		std::string& featureSearch,
+		std::map<std::string, bool>& categoryExpansionStates);
+
+	static void RenderRightColumn(
+		const std::vector<MenuFuncInfo>& menuList,
+		size_t selectedMenu);
+};
